% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{codeQueries}
\alias{codeQueries}
\title{Annotate a data frame of tokens with codes using Lucene-like search queries}
\usage{
codeQueries(tokens, queries, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  word.col = getOption("word.col", "word"), batchsize = 5,
  default.window = NA, condition_once = FALSE, indicator_filter = rep(T,
  nrow(tokens)), presorted = F, verbose = T)
}
\arguments{
\item{tokens}{a data frame of tokens containing columns for document id (), text position (position) and text string (column name can be specified in word.col, defaults to 'word').}

\item{queries}{a data frame containing the queries.}

\item{word.col}{a character string giving the name of the term string column}

\item{batchsize}{This function is faster if multiple queries are searched together, but too many queries (with too many tokens) at once can eat up memory or crash R. Try lowering batchsize in case of issues.}

\item{condition_once}{logical. If TRUE, then if an indicator satisfies its conditions once in an article, all indicators within that article are coded.}

\item{presorted}{The data has to be sorted on order(, position). If this is already the case, presorted can be set to TRUE to save time (which is usefull when testing many individual queries for large tokenlists)}
}
\value{
the annotated tokens data frame
}
\description{
Annotate a data frame of tokens with codes using Lucene-like search queries
}

