% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{codeTokens}
\alias{codeTokens}
\title{Annotate a data frame of tokens with codes using Lucene-like search queries}
\usage{
codeTokens(tokens, queries, text_var = "word", default.window = NA,
  indicator_filter = rep(T, nrow(tokens)), condition_once = FALSE,
  presorted = F)
}
\arguments{
\item{tokens}{a data frame of tokens containing columns for document id (doc_id), text position (position) and text string (column name can be specified in text_var, defaults to 'word').}

\item{queries}{a data frame containing the queries.}

\item{text_var}{a character string giving the name of the term string column}

\item{condition_once}{logical. If TRUE, then if an indicator satisfies its conditions once in an article, all indicators within that article are coded.}

\item{presorted}{The data has to be sorted on order(doc_id, position). If this is already the case, presorted can be set to TRUE to save time (which is usefull when testing many individual queries for large tokenlists)}
}
\value{
the annotated tokens data frame
}
\description{
Annotate a data frame of tokens with codes using Lucene-like search queries
}

