% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymatrix.r
\name{getQueryMatrix}
\alias{getQueryMatrix}
\title{Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit}
\usage{
getQueryMatrix(tokens, query_regex, aid_var = "aid", term_i_var = "id",
  term_var = "word", presorted = F, default.window = 25)
}
\arguments{
\item{article_filter}{}
}
\value{
a sparse matrix
}
\description{
Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit
}

