% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.r
\name{getTokenlist}
\alias{getTokenlist}
\title{Create a tokenlist data.frame}
\usage{
getTokenlist(x, doc_id = 1:length(x), removePunct = T, ...)
}
\arguments{
\item{...}{}
}
\description{
A tokenlist is a data.frame in which rows represent tokens (e.g., words)
}

