% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_query.r
\name{kwic}
\alias{kwic}
\title{Get keyword-in-context from a token list}
\usage{
kwic(tokens, hits, nwords = 10, text_var = "word", prettypaste = T)
}
\arguments{
\item{tokens}{a data frame of tokens containing columns for document id (doc_id), text position (position) and text string (column name can be specified in text_var, defaults to 'word').}

\item{nwords}{the number of words in front and after the keyword}

\item{text_var}{a character string giving the name of the term string column}
}
\value{
A data.frame with the keyword in context
}
\description{
Get keyword-in-context from a token list
}

