% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{reportKWIC}
\alias{reportKWIC}
\title{Title}
\usage{
reportKWIC(tokens, hits, nwords = 10, nsample = 10, random_sample = T,
  aid_var = "aid", term_i_var = "id", term_var = "word")
}
\arguments{
\item{term_var}{}
}
\description{
Title
}

