% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{reportQuery}
\alias{reportQuery}
\title{Title}
\usage{
reportQuery(tokens, hits, tokenfreq = T, keywordIC = T, kwic_nwords = 10,
  kwic_sample = 10, random_sample = T, aid_var = "aid",
  term_i_var = "id", term_var = "word")
}
\arguments{
\item{term_var}{}
}
\description{
Title
}

