% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{reportSummary}
\alias{reportSummary}
\title{Title}
\usage{
reportSummary(tokens, hits, hitcount = T, tokenfreq = T, keywordIC = T,
  kwic_nwords = 10, kwic_sample = 10, random_sample = T,
  aid_var = "aid", term_i_var = "id", term_var = "word")
}
\arguments{
\item{term_var}{}
}
\description{
Title
}

