% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{tokenGrepl}
\alias{tokenGrepl}
\title{a wrapper for grepl that takes multiple patterns. (and hopefully to optimize speed in the near future)}
\usage{
tokenGrepl(patterns, x, ignore.case = T, perl = F)
}
\arguments{
\item{perl}{}
}
\value{
a logical vector
}
\description{
a wrapper for grepl that takes multiple patterns. (and hopefully to optimize speed in the near future)
}

