% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querymatrix.r
\name{getQueryMatrix}
\alias{getQueryMatrix}
\title{Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit}
\usage{
getQueryMatrix(tokens, query_regex, text_var = "word", presorted = F,
  default.window = NA)
}
\arguments{
\item{default.window}{}
}
\value{
a sparse matrix
}
\description{
Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit
}

